feat: Update login component to handle inactive users
The login component has been updated to handle the case where the user is inactive. If the user is inactive, a message will be displayed indicating that the user no longer exists. This improves the user experience by providing clear feedback when attempting to log in with an inactive account.