feat: Add CollaboratorService and initialize data in IndexCollaboratorComponent
The code changes in index-collaborator.component.ts import the CollaboratorService and initialize the data in the IndexCollaboratorComponent. The CollaboratorService is used to make an API call to retrieve the list of collaborators, and the retrieved data is stored in the 'collaborators' variable. This feature enhances the application by populating the collaborator list on page load.