feat: Add getClients method to ClientService
This commit adds a new method called `getClients` to the `ClientService` in the `client.service.ts` file. The method makes an HTTP GET request to the `/getClients` endpoint with the provided `filtro` and `token` parameters. It sets the necessary headers and returns an Observable with the response data. This addition allows for retrieving clients based on a filter and authentication token.