feat: Add createdAt field to Collaborator and Client models
This commit adds the `createdAt` field to the `Collaborator` and `Client` models in the `Collaborator.js` and `Client.js` files, respectively. The `createdAt` field is of type `Date` and is set to the current date and time by default. This change ensures that the creation date of collaborators and clients is recorded and can be used for various purposes, such as sorting and filtering data.