feat: Add functionality to create client emails in Prospercción
This commit adds the functionality to create client emails in the Prospercción feature. It includes the creation of the `createClientEmailsProsperccion` function in the `prosperccionController.js` file, which handles the creation of client emails. It also introduces the `ClientEmails` model in the `ClientEmail.js` file, which represents client emails in the database. This enhancement allows users to send emails to clients and track their interactions within the Prospercción feature.